# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ECDAT Backend API (Module M7)",
    "description": "Enterprise Cryptographic Discovery & Analysis Tool (ECDAT) Backend API.\nCoordinates discovery scanning, Mosca quantum risk assessment, PQC recommendations, and CycloneDX 1.6 Cryptography Bill of Materials (CBOM) export.",
    "version": "1.0.0"
  },
  "paths": {
    "/scans": {
      "delete": {
        "tags": [
          "Scans"
        ],
        "summary": "Delete All Scans",
        "description": "Remove all saved history; reject the whole action while any scan is active.",
        "operationId": "delete_all_scans_scans_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Scans"
        ],
        "summary": "Create Scan",
        "description": "Kicks off an asynchronous cryptographic discovery scan.\nReturns immediately with a scanId and 'queued' status.",
        "operationId": "create_scan_scans_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Scans"
        ],
        "summary": "List Scans",
        "description": "Lists recent scans.",
        "operationId": "list_scans_scans_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScanResponse"
                  },
                  "title": "Response List Scans Scans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scans/{scan_id}": {
      "delete": {
        "tags": [
          "Scans"
        ],
        "summary": "Delete Scan",
        "description": "Remove a finished scan and its saved results. Input files are retained.",
        "operationId": "delete_scan_scans__scan_id__delete",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Scans"
        ],
        "summary": "Get Scan",
        "description": "Polls the status of an ongoing or completed scan.",
        "operationId": "get_scan_scans__scan_id__get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scans/upload": {
      "post": {
        "tags": [
          "Scans"
        ],
        "summary": "Upload Scan",
        "description": "Persist a single source file, ZIP workspace, or image SBOM before queueing.",
        "operationId": "upload_scan_scans_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_scan_scans_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_scan_scans_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "sourceType": {
            "type": "string",
            "title": "Sourcetype",
            "default": "upload"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_scan_scans_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScanCreate": {
        "properties": {
          "sourceType": {
            "type": "string",
            "enum": [
              "git",
              "upload",
              "image",
              "path"
            ],
            "title": "Sourcetype",
            "description": "Scan input source type",
            "default": "path"
          },
          "target": {
            "type": "string",
            "title": "Target",
            "description": "Target repository URL, local path, container image reference, or uploaded file"
          },
          "complianceTarget": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "NIST-general",
                  "CNSA2.0"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Compliancetarget",
            "default": "NIST-general"
          },
          "threatTimelineOverride": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 50.0,
                "minimum": 0.5
              },
              {
                "type": "null"
              }
            ],
            "title": "Threattimelineoverride",
            "description": "Optional custom Z (years until CRQC) override"
          }
        },
        "type": "object",
        "required": [
          "target"
        ],
        "title": "ScanCreate"
      },
      "ScanResponse": {
        "properties": {
          "scanId": {
            "type": "string",
            "title": "Scanid"
          },
          "sourceType": {
            "type": "string",
            "title": "Sourcetype"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "assetCount": {
            "type": "integer",
            "title": "Assetcount"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "completedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "scanId",
          "sourceType",
          "target",
          "status",
          "progress",
          "assetCount",
          "createdAt"
        ],
        "title": "ScanResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}